Adicionado modelo de Inteligência Artificial

+++ cardapio-sesc/backend/index.js
@@ -1,19 +1,25 @@
 import { DynamoDBClient, PutItemCommand, QueryCommand } from "@aws-sdk/client-dynamodb"
 import { parse } from "node-html-parser"
 import { createWorker } from "tesseract.js"
-import {Jimp} from "jimp"
+import { Jimp } from "jimp"
+import OpenAI from 'openai'
 
 const dynamo = new DynamoDBClient({ region: "sa-east-1" })
 
+const client = new OpenAI({
+  apiKey: process.env['OPENAI_API_KEY'],
+});
+
+
 export async function handler(event) {
   //CORS preflight
-  if (event.requestContext?.http?.method === "OPTIONS") {
-    return {
-      statusCode: 200,
-      headers: corsHeaders(),
-      body: "",
-    }
-  }
+  // if (event.requestContext?.http?.method === "OPTIONS") {
+  //   return {
+  //     statusCode: 200,
+  //     headers: corsHeaders(),
+  //     body: "",
+  //   }
+  // }
 
   try {
     await processMenu()
@@ -49,9 +55,9 @@ async function processMenu() {
 
   const html = await response.text()
   const root = parse(html)
- const img = root.querySelector(".entry-content img")
+  const img = root.querySelector(".entry-content img")
 
-  if(img == null){
+  if (img == null) {
     throw new Error("Mudança no seletor da classe que contem a img. Corrija o seletor")
   }
   const src = img.getAttribute("src")
@@ -91,9 +97,10 @@ async function processMenu() {
       x: crops[i],
       y: marginTop,
     })
-    
+
     // const dateCrop = image.clone().crop(crops[i], dateMarginTop, filledWidth, dateHeight)
-    const dateCrop = image.clone().crop({ h: dateHeight,
+    const dateCrop = image.clone().crop({
+      h: dateHeight,
       w: filledWidth,
       x: crops[i],
       y: dateMarginTop,
@@ -142,20 +149,37 @@ async function processMenu() {
 
 async function getText(cropped, croppedData) {
   const worker = await createWorker("eng")
-// // //Descomentar para ajudar na depuração
-    // const file = './' + "teste" + '.png'
-    // await cropped.writeAsync(file)
-    // console.log('Escrito arquivo ' + file)
+  // // //Descomentar para ajudar na depuração
+  // const file = './' + "teste" + '.png'
+  // await cropped.writeAsync(file)
+  // console.log('Escrito arquivo ' + file)
   const buffer = await cropped.getBuffer("image/png")
   const bufferData = await croppedData.getBuffer("image/png")
 
-  const texto = await worker.recognize(buffer)
+  let textoReconhecido = await worker.recognize(buffer)
   const data = await worker.recognize(bufferData)
+  let texto = textoReconhecido.data.text
+
+  const response = await client.chat.completions.create({
+    model: 'gpt-4.1-mini',
+    messages: [
+      { role: 'system', content: 'Você irá ajudar a refinar textos extraídos de um OCR. É necessário corrigir a grafia e formato do texto. Os textos são opções de refeições, use isso para preencher melhor as palavras. O formato da responsta deve ser exclusivamente JSON com uma única propriedade "text". A resposta com qualquer outro formato ou propriedades causará erros.' },
+      { role: 'user', content: texto }
+    ],
+    response_format: { type: 'json_object' }
+  });
+
+  texto = JSON.parse(response.choices[0].message.content)
+  if (typeof texto !== 'object' || typeof texto.text != 'string') {
+    console.error(`O chatgpt respondeu num formato inválido: ` + JSON.stringify(texto))
+    const outraTentativa = await getText(cropped, croppedData)
+    return outraTentativa
+  }
 
   await worker.terminate()
 
   return {
-    text: texto.data.text.trim(),
+    text: texto.text,
     date: data.data.text.trim(),
   }
 }
@@ -185,4 +209,4 @@ async function verifyVersion(date) {
   const data = await dynamo.send(new QueryCommand(params))
   return data.Items ?? []
 }
-//handler()
+handler()
